requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-adk>=1.0.0
aiohttp>=3.9.0,<3.10.0